pyspark-etl-project/etl/requirements.txt
pyspark==3.5.0
boto3==1.30.0
PyMySQL==1.0.2
pyyaml==6.0